Use `setup_test_component_platform` helper for time entity component tests instead of `hass.components`
